Force float32 dtype when loading noobai_xl_vpred pipeline to CPU
The model has float16 weights in the text encoder's text_projection
layer. Without explicitly passing dtype=torch.float32 to pipe.to(),
these remain as float16 and cause a dtype mismatch during encoding.

+++ noobai_xl_vpred/pytorch/src/model_utils.py
@@ -26,7 +26,7 @@ def load_pipe(variant):
     pipe = StableDiffusionXLPipeline.from_pretrained(variant, torch_dtype=torch.float32)
     modules = [pipe.text_encoder, pipe.unet, pipe.text_encoder_2, pipe.vae]
 
-    pipe.to("cpu")
+    pipe.to("cpu", dtype=torch.float32)
 
     for module in modules:
         module.eval()
